feat: use file reference to construct readOp objects
There is no need to extract (and the copy) file name and object size
from file reference to construct read object, we can use file reference
directly.
